refactor: wallet 예외 패키지 구조 정리
