+++ modi/account-service/src/main/java/com/coc/modi/account/deposit/application/DepositService.java
@@ -9,8 +9,8 @@ import com.coc.modi.account.deposit.domain.PgDepositRepository;
 import com.coc.modi.account.deposit.infrastructure.client.TossPaymentsClient;
 import com.coc.modi.account.deposit.infrastructure.client.dto.TossPaymentApprovalResponse;
 import com.coc.modi.account.deposit.infrastructure.client.dto.TossPaymentCancelResponse;
-import com.coc.modi.account.exception.AccountException;
-import com.coc.modi.account.exception.AccountTransactionNotFoundException;
+import com.coc.modi.account.wallet.exception.AccountException;
+import com.coc.modi.account.wallet.exception.AccountTransactionNotFoundException;
 import com.coc.modi.account.wallet.application.WalletCommandService;
 import com.coc.modi.account.wallet.application.dto.WalletTransactionCommand;
 import com.coc.modi.common.ErrorCode;
