+++ modi/account-service/src/main/java/com/coc/modi/account/deposit/domain/PgDeposit.java
@@ -1,6 +1,6 @@
 package com.coc.modi.account.deposit.domain;
 
-import com.coc.modi.account.exception.AccountException;
+import com.coc.modi.account.wallet.exception.AccountException;
 import com.coc.modi.common.BaseEntity;
 import com.coc.modi.common.ErrorCode;
