+++ modi/account-service/src/main/java/com/coc/modi/account/wallet/application/MemberWalletService.java
@@ -1,6 +1,6 @@
 package com.coc.modi.account.wallet.application;
 
-import com.coc.modi.account.exception.AccountNotFoundException;
+import com.coc.modi.account.wallet.exception.AccountNotFoundException;
 import com.coc.modi.account.wallet.application.dto.MemberWalletResponse;
 import com.coc.modi.account.wallet.application.dto.WalletTransactionResponse;
 import com.coc.modi.account.wallet.domain.MemberWallet;
