+++ modi/account-service/src/main/java/com/coc/modi/account/wallet/exception/AccountException.java
@@ -1,4 +1,4 @@
-package com.coc.modi.account.exception;
+package com.coc.modi.account.wallet.exception;
 
 import com.coc.modi.common.BaseException;
 import com.coc.modi.common.ErrorCode;
